ShowsPopularFragment: extract common code.
